pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "openai>=1.99.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
